+++ src/features/question-fivesec/answer/FivesecCountdownPhase.tsx
@@ -27,10 +27,11 @@ export function FivesecCountdownPhase({ remaining, imageUrl, ratio, hideCountdow
         categoryLabel="5초 테스트"
         title="5초간 아래 사진에 집중해주세요"
       />
-      <div className="flex-1 px-5 pt-4">
+      <div className="flex-1 px-5 pt-4 flex justify-center">
         <div
-          className="relative w-full overflow-hidden rounded-2xl"
+          className="relative overflow-hidden rounded-2xl"
           style={{
+            width: `min(100%, calc((100dvh - 200px) * ${RATIO_TO_CSS[ratio]}))`,
             aspectRatio: RATIO_TO_CSS[ratio],
             boxShadow: "inset 0 0 0 1px rgba(2,32,71,0.05)",
           }}
