+++ src/features/question-fivesec/answer/FivesecPreviewPhase.tsx
@@ -18,11 +18,8 @@ interface Props {
 export function FivesecPreviewPhase({ ratio, onStart }: Props) {
   return (
     <div className="flex flex-col flex-1 bg-white">
-      <QuestionHeader
-        categoryLabel="5초 테스트"
-        title="5초간 아래 사진에 집중해주세요"
-      />
-      <div className="flex-1 px-5 pt-4">
+      <QuestionHeader categoryLabel="5초 테스트" title={`5초간\n아래 사진에 집중해주세요`} />
+      <div className="flex-1 px-5 pt-4 flex justify-center">
         <motion.div
           role="button"
           onClick={onStart}
@@ -30,7 +27,7 @@ export function FivesecPreviewPhase({ ratio, onStart }: Props) {
           animate={{ scale: 1, opacity: 1 }}
           transition={{ duration: 0.25, ease: "easeOut" }}
           style={{
-            width: "100%",
+            width: `min(100%, calc((100dvh - 200px) * ${RATIO_TO_CSS[ratio]}))`,
             aspectRatio: RATIO_TO_CSS[ratio],
             backgroundColor: adaptive.grey100,
             borderRadius: 16,
